// api/src/modules/database/constants/schema-names.ts
// Dont change values, they describe the names of the schemas in database
export enum SchemaNames {
  ACCESS_CARD = 'ACCESS_CARD',
  AGENCY = 'AGENCY',
  CLIENT = 'CLIENT',
  DOER = 'WORKER',
  ENTRY_POINT = 'ENTRY_POINT',
  JOB_BOARD = 'JOB_BOARD',
  MISSION = 'MISSION',
  QUESTION = 'QUESTION',
  QUESTION_SHEET = 'QUESTION_SHEET',
  QUIZZ = 'QUIZZ',
  QUIZZ_SHEET = 'QUIZZ_SHEET',
}